added exercise detection code, real time demo with threading
