+++ .gitignore
@@ -19,8 +19,11 @@ __pycache__/
 database/.DS_Store
 .idea
 Metamotion/Mobitrack_scripts/.metawear
+env/
 
 # Data locations
 data/Jan28ClassifierResults
 data/Jan27ClassifierResults
+data
 database/node_modules/
+
