+++ database/requirements.txt
@@ -1,6 +1,19 @@
-django==2.1.3
+certifi==2018.11.29
+chardet==3.0.4
+cycler==0.10.0
+Django==2.1.3
 djangorestframework==3.9.1
+idna==2.8
+kiwisolver==1.0.1
+matplotlib==3.0.2
+metawear==0.7.0
+mysql-connector-python==8.0.13
+numpy==1.16.1
 protobuf==3.6.1
+pyparsing==2.3.1
+python-dateutil==2.8.0
 pytz==2018.7
+requests==2.21.0
 six==1.11.0
-mysql-connector-python == 8.0.13
+urllib3==1.24.1
+warble==1.1.0
