+++ processing/old/Jan27_ML_segmentation/formatDataForClassifier.m
