+++ processing/old/Jan27_ML_segmentation/formatDataForTraining.m
