+++ processing/old/Jan27_ML_segmentation/loadDataFromTxtFile.m
