+++ processing/old/Jan27_ML_segmentation/plotquick.m
