+++ processing/old/extractSegments.m
