+++ processing/old/getSVM.m
