+++ processing/old/readcsvfile.m
