+++ processing/old/segmentData.m
