+++ processing/old/trainClassifier.m
