+++ processing/old/trainClassifierWorkflow.m
