+++ processing/old/trainingData/testing_nov17.20-0 striated.mat
