+++ processing/old/trainingData/training_nov17.20-0 striated.mat
